requirements.txt
streamlit
anthropic
python-dotenv
beautifulsoup4
requests
yt-dlp
openai-whisper
scrapy
selenium
webdriver-manager
schedule
plotly
python-docx


streamlit-option-menu

